+++ src/javascript/API.js
@@ -445,6 +445,13 @@ const graviton = {
     }else{
       return true
     }
+  },
+  resizeTerminals(){
+    for(i=0;i<editor_screens.length;i++){
+      if(editor_screens[i].terminal !=undefined){
+        fit.fit(editor_screens[i].terminal.xterm);
+      }
+    }
   }
 }
 
@@ -935,8 +942,17 @@ const commanders = {
             theme:{
               background:graviton.getCurrentTheme().Colors["editor-background-color"],
               foreground:graviton.getCurrentTheme().Colors["white-black"]
+            },
+            onResize: function(){
+              console.log("pro");
             }
           });
+          //
+          console.log(randomID);
+          document.getElementById("xterm"+randomID+"_commander").addEventListener("resize",function(e){
+            fit.fit(current_screen.terminal.xterm)
+            console.log("you are pro!");
+          })
           xterm.open(document.getElementById("xterm"+randomID+"_commander"));  
           xterm.on('data', (data) => {
             ptyProcess.write(data);
@@ -944,6 +960,7 @@ const commanders = {
           ptyProcess.on('data', function (data) {
             xterm.write(data);
           });
+          xterm
           for(i=0;i<editor_screens.length;i++){
             if(editor_screens[i].id==current_screen.id){
               editor_screens[i].terminal = {id:"xterm"+randomID,xterm:xterm};
@@ -1087,3 +1104,4 @@ const screens = {
     }
   }
 }
+window.onresize = function(){console.log("pro"); graviton.resizeTerminals();}